chore(hog-charts): move LineChart into its own folder
Every other chart type follows the documented `charts/<name>/<Name>.tsx` folder layout (`BarChart/`, `BoxPlot/`, `PieChart/`, `Sparkline/`, `TimeSeriesBarChart/`, `TimeSeriesLineChart/`). LineChart was the lone holdout sitting flat at `charts/LineChart.tsx`.

Move the three files (`LineChart.tsx`, `LineChart.test.tsx`, `LineChart.stories.tsx`) into `charts/LineChart/`, bump the relative imports inside them from `../` to `../../`, and update the four import sites (the library barrel and two overlay stories). All renames preserved via `git mv` so blame still works.

Generated-By: PostHog Code
Task-Id: 014c6695-6ee4-44fc-9f38-34018338582e
