+++ frontend/src/lib/hog-charts/charts/LineChart/LineChart.stories.tsx
@@ -1,11 +1,11 @@
 import { Meta, StoryObj } from '@storybook/react'
 
-import type { LineChartConfig, Series } from '../core/types'
-import { DEFAULT_Y_AXIS_ID } from '../core/types'
-import { ReferenceLine } from '../overlays/ReferenceLine'
-import { ValueLabels } from '../overlays/ValueLabels'
-import { playHoverAtFraction, Stage, useReactiveTheme } from '../story-helpers'
-import { ciRanges, trendLine } from '../utils/statistics'
+import type { LineChartConfig, Series } from '../../core/types'
+import { DEFAULT_Y_AXIS_ID } from '../../core/types'
+import { ReferenceLine } from '../../overlays/ReferenceLine'
+import { ValueLabels } from '../../overlays/ValueLabels'
+import { playHoverAtFraction, Stage, useReactiveTheme } from '../../story-helpers'
+import { ciRanges, trendLine } from '../../utils/statistics'
 import { LineChart } from './LineChart'
 
 const DAYS = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun']
