+++ frontend/src/lib/hog-charts/index.ts
@@ -1,8 +1,8 @@
 // Components
 export { BarChart } from './charts/BarChart/BarChart'
 export type { BarChartProps } from './charts/BarChart/BarChart'
-export { LineChart } from './charts/LineChart'
-export type { LineChartProps } from './charts/LineChart'
+export { LineChart } from './charts/LineChart/LineChart'
+export type { LineChartProps } from './charts/LineChart/LineChart'
 export { TimeSeriesLineChart } from './charts/TimeSeriesLineChart/TimeSeriesLineChart'
 export type {
     ConfidenceIntervalConfig,
